Add files via upload
This is project that I have worked on in my semester break at SRGEC. This is one of the project that I built with the help of chat GPT 4.0. This the the first project of mine that I created with the help of AI . and This project helped me to gain intrest on prompt engineering domain. But as known the top level of web development is crafted by me and the foundation level labour is done by chatgpt.